requirements.txt
certifi==2019.11.28
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
olefile==0.46
Pillow==7.0.0
Werkzeug==0.16.0
wincertstore==0.2
azure-mgmt-resource==0.30.0rc5